+++ lib/tz_world/backend/ets_with_index_cache.ex
@@ -190,10 +190,12 @@ defmodule TzWorld.Backend.EtsWithIndexCache do
     end)
   end
 
+  @doc false
   def get_index_cache do
     :ets.select(__MODULE__, index_spec())
   end
 
+  @doc false
   def index_spec do
     [{{{:"$1", :"$2", :"$3", :"$4"}, :"$5"}, [], [{{:"$1", :"$2", :"$3", :"$4"}}]}]
   end